Use SUPPORTS_SYSTEM_TEXT_JSON instead of NET6_0_OR_GREATER. Move net6 to platform folder to conditionally compile.
