+++ src/client/Microsoft.Identity.Client/Cache/Adal/AdalResultWrapper.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using Microsoft.Identity.Client.Utils;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using Microsoft.Identity.Client.Platforms.net6;
 #else
 using Microsoft.Identity.Json;
@@ -15,12 +15,12 @@ namespace Microsoft.Identity.Client.Cache
     internal class AdalResultWrapper
     {
 
-#if !NET6_0_OR_GREATER
+#if !SUPPORTS_SYSTEM_TEXT_JSON
         [JsonProperty]
 #endif
         public AdalResult Result { get; set; }
 
-#if !NET6_0_OR_GREATER
+#if !SUPPORTS_SYSTEM_TEXT_JSON
         [JsonProperty]
 #endif
         public string RawClientInfo { get; set; }
@@ -28,7 +28,7 @@ namespace Microsoft.Identity.Client.Cache
         /// <summary>
         /// Gets the Refresh Token associated with the requested Access Token. Note: not all operations will return a Refresh Token.
         /// </summary>
-#if !NET6_0_OR_GREATER
+#if !SUPPORTS_SYSTEM_TEXT_JSON
         [JsonProperty]
 #endif
         public string RefreshToken { get; set; }
@@ -40,7 +40,7 @@ namespace Microsoft.Identity.Client.Cache
 
         // This is only needed for AcquireTokenByAuthorizationCode in which parameter resource is optional and we need
         // to get it from the STS response.
-#if !NET6_0_OR_GREATER
+#if !SUPPORTS_SYSTEM_TEXT_JSON
         [JsonProperty]
 #endif
         internal string ResourceInResponse { get; set; }
@@ -63,7 +63,7 @@ namespace Microsoft.Identity.Client.Cache
             return JsonHelper.SerializeToJson(this);
         }
 
-#if !NET6_0_OR_GREATER
+#if !SUPPORTS_SYSTEM_TEXT_JSON
         [JsonProperty]
 #endif
         public string UserAssertionHash { get; set; }
