+++ src/client/Microsoft.Identity.Client/Cache/ITokenCacheSerializable.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using JToken = System.Text.Json.Nodes.JsonNode;
 #else
 using Microsoft.Identity.Json.Linq;
