+++ src/client/Microsoft.Identity.Client/Cache/TokenCacheDictionarySerializer.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.Utils;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using JToken = System.Text.Json.Nodes.JsonNode;
 #else
 using Microsoft.Identity.Json.Linq;
