+++ src/client/Microsoft.Identity.Client/Cache/TokenCacheJsonSerializer.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.Utils;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using JToken = System.Text.Json.Nodes.JsonNode;
 #else
 using Microsoft.Identity.Json.Linq;
