+++ src/client/Microsoft.Identity.Client/Instance/Validation/AdfsWebFingerResponse.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 using Microsoft.Identity.Client.OAuth2;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using Microsoft.Identity.Client.Platforms.net6;
 using System.Text.Json.Nodes;
 using JsonProperty = System.Text.Json.Serialization.JsonPropertyNameAttribute;
