+++ src/client/Microsoft.Identity.Client/Internal/DeviceCodeResponse.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Microsoft.Identity.Client.OAuth2;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using Microsoft.Identity.Client.Platforms.net6;
 using JsonProperty = System.Text.Json.Serialization.JsonPropertyNameAttribute;
 #else
