+++ src/client/Microsoft.Identity.Client/MsalServiceException.cs
@@ -5,7 +5,7 @@ using System;
 using System.Globalization;
 using System.Net.Http.Headers;
 using Microsoft.Identity.Client.Utils;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using System.Text.Json.Serialization;
 using JObject = System.Text.Json.Nodes.JsonObject;
 #else
@@ -164,7 +164,7 @@ namespace Microsoft.Identity.Client
         /// </list>
         /// For more details see https://aka.ms/msal-net-claim-challenge
         /// </summary>
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
         [JsonInclude]
 #endif
         public string Claims { get; internal set; }
