+++ tests/Microsoft.Identity.Test.Unit/CacheTests/CacheSerializationTests.cs
@@ -14,7 +14,7 @@ using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.Utils;
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using JObject = System.Text.Json.Nodes.JsonObject;
@@ -45,7 +45,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             StorageJsonKeys.FamilyId
         };
 
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
         private static readonly JsonDocumentOptions _documentOptions = new JsonDocumentOptions
         {
             AllowTrailingCommas = true
@@ -538,7 +538,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             await (tokenCache as ITokenCacheInternal).OnAfterAccessAsync(notification).ConfigureAwait(false);
             (tokenCache as ITokenCacheInternal).Accessor.AssertItemCount(5, 4, 3, 3, 3);
 
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
             var finalJson = JsonNode.Parse(Encoding.UTF8.GetString(cache), documentOptions: _documentOptions).AsObject();
 
             var originalJson = JsonNode.Parse(jsonContent, documentOptions: _documentOptions).AsObject();
@@ -562,7 +562,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             byte[] bytes = s1.Serialize(null);
             string actualJson = new UTF8Encoding().GetString(bytes);
 
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
             Assert.IsTrue(DeepEquals(
                 JsonNode.Parse(actualJson, documentOptions: _documentOptions).AsObject(),
                 JsonNode.Parse(expectedJson, documentOptions: _documentOptions).AsObject()));
@@ -578,7 +578,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             // serialize again to detect errors that come from deserialization
             byte[] bytes2 = s2.Serialize(null);
             string actualJson2 = new UTF8Encoding().GetString(bytes2);
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
             Assert.IsTrue(DeepEquals(
                     JsonNode.Parse(actualJson2, documentOptions: _documentOptions).AsObject(),
                 JsonNode.Parse(expectedJson, documentOptions: _documentOptions).AsObject()));
@@ -636,7 +636,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             // serialize again to detect errors that come from deserialization
             byte[] bytes2 = s1.Serialize(null);
             string actualJson2 = new UTF8Encoding().GetString(bytes2);
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
             Assert.IsTrue(DeepEquals(
                 JsonNode.Parse(actualJson2, documentOptions: _documentOptions).AsObject(),
                 JsonNode.Parse(expectedJson, documentOptions: _documentOptions).AsObject()));
@@ -938,7 +938,7 @@ namespace Microsoft.Identity.Test.Unit.CacheTests
             CoreAssert.AssertDictionariesAreEqual(expected.WamAccountIds, actual.WamAccountIds, StringComparer.Ordinal);
         }
 
-#if NET6_0_OR_GREATER
+#if SUPPORTS_SYSTEM_TEXT_JSON
         private bool DeepEquals(JsonNode a, JsonNode b)
         {
             if (a == null && b == null)
